Prefer current to simple when pushing branches
If a branch only has pushRemote set, then we can't simply use `git
push`:

    $ git push
    fatal: The current branch example_branch has no upstream branch.
    To push the current branch and set the remote as upstream, use

        git push --set-upstream origin example_branch

This is because `push.default = simple` assumes a 'central workflow':
https://git-scm.com/docs/git-config#git-config-pushdefault and
pushRemote signals to git that we don't have a central workflow.

`current` supports both central and non-central workflows, so we can
now use `git push` directly. Since magit makes it easy to set
pushRemote, it was easy to get into a state where we can push from
magit but not from a shell.